samples: add proactive reclamation DAMON sample module
Add a sample DAMON application kernel module, namely damon_sample_prcl.
It is same to the one of the example kernel modules that presented at
kernel summit 2021 with 10 minutes live coding[1] but updated for the
latest DAMON interface.

The sample module receives a pid of a process to monitor the access,
monitors the access to the virtual address space of the process, find
regions that not accessed and proactively reclaim those using DAMOS.

[1] https://linuxplumbersconf.org/event/11/contributions/984/

Link: https://patch.msgid.link/20231205022858.1540-4-sj@kernel.org
Signed-off-by: SeongJae Park <sj@kernel.org>