==== docs improvement for mm ====
Subject: Docs/mm: rename legacy docs to unsorted docs

The work for changing the memory mangement documentation outline has
started from 2022.  For that, old documents have placed under new
chapter, "Legacy Documentations".  The plan is to eventually move all
old documentations into new outline, while asking new documents to be
added on the new outline from the beginning.

However, it is unclear where on the new outline each document should be
placed for some cases.  Meanwhile, the name ("legacy") makes readers
feel the documents under the chapter might be outdated or not actively
maintained.  The absence of clear criteria for new outline also makes
new documents difficult to find where those should be placed.  A new
document was put on the bottom of the new outline recently, apparently
not based on a clear guideline but just because it is the bottom of the
list.

Rename the "Legagy Documentation" chapter to "Unsorted Documentations"
with minor fixups.

Signed-off-by: SeongJae Park <sj@kernel.org>