mm/damon/core: handle zero schemes apply interval
DAMON's logics to determine if this is the time to apply damos schemes
assumes next_apply_sis is always set larger than current
passed_sample_intervals.  And therefore continuously incrementing
passed_sample_intervals will make it reaches to the next_apply_sis in
future.  The logic therefore checks if this is exactly the time, and
update next_apply_sis in the case.

If schemes apply interval is set as zero, however, next_apply_sis is set
same to current passed_sample_intervals, respectively.  And
passed_sample_intervals is incremented before doing the next_apply_sis
check.  Hence, next_apply_sis is larger than next_apply_sis, and the
logic says it is not the time to apply schemes and update
next_apply_sis.  In other words, DAMON stops applying schemes.

Based on the documents and the common sense, a reasonable behavior for
such inputs would be applying schemes for every sampling interval.
Handle the case by removing the assumption.

Fixes: 42f994b71404 ("mm/damon/core: implement scheme-specific apply interval")
Signed-off-by: SeongJae Park <sj@kernel.org>