mm/damon/core: implement damon_call()
Introduce a new DAMON core API function, damon_call().  It aims to
replace some damon_callback usages that access damon_ctx of ongoing
kdamond with additional synchronizations.  It receives a function
pointer, let the parallel kdamond invokes the function, and returns
after the invocation is finished, or canceled due to some races.

kdamond invokes the function inside the main loop after sampling is
done.  If it is deactivated by DAMOS watermarks or already out of the
main loop, mark the request as canceled so that damon_call() can wakeup
and return.

Signed-off-by: SeongJae Park <sj@kernel.org>