mm/damon/core: implement damos_walk()
Introduce a new core layer interface, damos_walk().  It aims to replace
some damon_callback usages that access DAMOS schemes applied regions of
ongoing kdamond with additional synchronizations.  It receives a
function pointer and asks kdamond to invoke it for any region that it will
apply any DAMOS action within one scheme apply interval for every scheme
of it.  The function further waits until the kdamond finishes the
invocations for every scheme, or cancels the request, and returns.

The kdamond invokes the function as requested within the main loop.  If
it is deactivated by DAMOS watermarks or going out of the main loop, it
marks the request as canceled, so that damos_walk() can wakeup and
return.

Signed-off-by: SeongJae Park <sj@kernel.org>