mm/damon/core: invoke damos_walk_control->walk_fn() after applying action
damos_walk_control->walk_fn() is called just before a DAMOS action is
applied to a region.  It can be useful if some users want to make some
changes to the region before the scheme's action is applied, but there
is no such real usages.

On the other hand, calling it after the DAMOS action is applied can give
a chance to investigate what happened during applying the action to the
region.  Actually, followup commits will make and pass such information
to walk_fn().  Invoke walk_fn() after applying the action.

Signed-off-by: SeongJae Park <sj@kernel.org>