Docs/mm/damon/design: add monitoring parameters tuning guide
DAMON monitoring parameters including sampling and aggregation intervals
require tuning based on the given workloads.  However, there is no
central guide.  This resulted in a few problem reports[1].  Add a guide
on the design document that might imperfect, but the community can
gather the efforts to make be more perfect.

[1] https://lore.kernel.org/20241202175459.2005526-1-sj@kernel.org

Signed-off-by: SeongJae Park <sj@kernel.org>