mm/damon: add default allow/reject behavior fields to struct damos
It is more intuitive to decide the default filtering stage allow/reject
behavior as opposite to the last filter's behavior.  The decision should
be made separately for core and operations layers' filtering stages,
since last core layer-handled filter is not really a last filter if
there are operations layer handling filters.

Keeping separate decisions for the two categories can make the logic
simpler.  Add fields for string the two decisions.

Signed-off-by: SeongJae Park <sj@kernel.org>