mm/damon/Kconfig: defaults VADDR, PADDR, and SYSFS to DAMON
It is suggested to enable DAMON_VADDR, DAMON_PADDR and DAMON_SYSFS
configurations together with DAMON, since those are essential parts of
DAMON for optimum and usual usages.  Because those need to be enabled
one by one, and there are other test-purpose or non-essential
configurations, it is easy to be confused and make mistakes at setup.
Make the essential configuration defaults to CONFIG_DAMON, so that those
can be enabled by default with single change.

Signed-off-by: SeongJae Park <sj@kernel.org>